API: update the generated code

+++ TdLib.Api/Functions/DownloadFile.cs
@@ -34,7 +34,7 @@ namespace TdLib
             public int FileId { get; set; }
 
             /// <summary>
-            /// Priority of the download (1-32). The higher the priority, the earlier the file will be downloaded. If the priorities of two files are equal, then the last one for which downloadFile was called will be downloaded first
+            /// Priority of the download (1-32). The higher the priority, the earlier the file will be downloaded. If the priorities of two files are equal, then the last one for which downloadFile/addFileToDownloads was called will be downloaded first
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("priority")]
@@ -45,17 +45,17 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("offset")]
-            public int Offset { get; set; }
+            public long Offset { get; set; }
 
             /// <summary>
             /// Number of bytes which need to be downloaded starting from the "offset" position before the download will automatically be canceled; use 0 to download without a limit
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("limit")]
-            public int Limit { get; set; }
+            public long Limit { get; set; }
 
             /// <summary>
-            /// If false, this request returns file state just after the download has been started. If true, this request returns file state only after
+            /// Pass true to return response only after the file download has succeeded, has failed, has been canceled, or a new downloadFile request with different offset/limit parameters was sent; pass false to return file state immediately, just after the download has been started
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("synchronous")]
@@ -66,7 +66,7 @@ namespace TdLib
         /// Downloads a file from the cloud. Download progress and completion of the download will be notified through updateFile updates
         /// </summary>
         public static Task<File> DownloadFileAsync(
-            this Client client, int fileId = default, int priority = default, int offset = default, int limit = default, bool synchronous = default)
+            this Client client, int fileId = default, int priority = default, long offset = default, long limit = default, bool synchronous = default)
         {
             return client.ExecuteAsync(new DownloadFile
             {
